Annotates with @VisibleForTesting. Makes the Helper functions package-private. Also adds a few tests for `getDynamicMessageFromGenericRecord()` method.Adds support for serialization of generic records to Proto Rows. Moves Descriptor to AvroToProtoSerializer. Removes Utils Class and its tests and move to AvroToProtoSerializer.
