+++ flink-connector-bigquery/src/main/java/com/google/cloud/flink/bigquery/sink/serializer/AvroToProtoSerializer.java
@@ -16,7 +16,11 @@
 
 package com.google.cloud.flink.bigquery.sink.serializer;
 
+import org.apache.flink.annotation.VisibleForTesting;
+
 import com.google.api.client.util.Preconditions;
+import com.google.cloud.bigquery.storage.v1.BigDecimalByteStringEncoder;
+import com.google.cloud.bigquery.storage.v1.CivilTimeEncoder;
 import com.google.cloud.flink.bigquery.sink.exceptions.BigQuerySerializationException;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.Descriptors.Descriptor;
@@ -27,13 +31,25 @@ import org.apache.avro.LogicalTypes;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
+import org.apache.avro.util.Utf8;
+import org.apache.commons.lang3.ArrayUtils;
+import org.joda.time.Days;
+import org.joda.time.Instant;
+import org.joda.time.ReadableInstant;
+import org.json.JSONException;
+import org.json.JSONObject;
 
 import javax.annotation.Nullable;
 
+import java.math.BigDecimal;
 import java.nio.ByteBuffer;
+import java.sql.Timestamp;
+import java.time.LocalDate;
 import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
 import java.util.function.UnaryOperator;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
@@ -41,7 +57,7 @@ import java.util.stream.StreamSupport;
 /** Serializer for converting Avro's {@link GenericRecord} to BigQuery proto. */
 public class AvroToProtoSerializer implements BigQueryProtoSerializer<GenericRecord> {
 
-    private final BigQuerySchemaProvider bigQuerySchemaProvider;
+    private Descriptor descriptor;
 
     private static final Map<Schema.Type, UnaryOperator<Object>> PRIMITIVE_TYPE_ENCODERS;
 
@@ -69,21 +85,22 @@ public class AvroToProtoSerializer implements BigQueryProtoSerializer<GenericRec
     }
 
     /**
-     * Constructor for the Serializer.
+     * Initializer for the Serializer.
      *
      * @param bigQuerySchemaProvider BigQuerySchemaProvider for the Sink Table ({@link
      *     BigQuerySchemaProvider} object)
      */
-    public AvroToProtoSerializer(BigQuerySchemaProvider bigQuerySchemaProvider) {
-        this.bigQuerySchemaProvider = bigQuerySchemaProvider;
+    public void init(BigQuerySchemaProvider bigQuerySchemaProvider) {
+        Preconditions.checkNotNull(
+                bigQuerySchemaProvider,
+                "BigQuerySchemaProvider not initialized before initializing Serializer.");
+        this.descriptor = bigQuerySchemaProvider.getDescriptor();
     }
 
     @Override
     public ByteString serialize(GenericRecord record) throws BigQuerySerializationException {
         try {
-            return getDynamicMessageFromGenericRecord(
-                            record, this.bigQuerySchemaProvider.getDescriptor())
-                    .toByteString();
+            return getDynamicMessageFromGenericRecord(record, this.descriptor).toByteString();
         } catch (Exception e) {
             throw new BigQuerySerializationException(e.getMessage(), e);
         }
@@ -202,34 +219,203 @@ public class AvroToProtoSerializer implements BigQueryProtoSerializer<GenericRec
      */
     private static UnaryOperator<Object> getLogicalEncoder(String logicalTypeString) {
         Map<String, UnaryOperator<Object>> mapping = new HashMap<>();
-        mapping.put(LogicalTypes.date().getName(), AvroToProtoSerializerUtils::convertDate);
-        mapping.put(
-                LogicalTypes.decimal(1).getName(), AvroToProtoSerializerUtils::convertBigDecimal);
+        mapping.put(LogicalTypes.date().getName(), AvroToProtoSerializer::convertDate);
+        mapping.put(LogicalTypes.decimal(1).getName(), AvroToProtoSerializer::convertBigDecimal);
         mapping.put(
                 LogicalTypes.timestampMicros().getName(),
-                value ->
-                        AvroToProtoSerializerUtils.convertTimestamp(
-                                value, true, "Timestamp(micros/millis)"));
+                value -> convertTimestamp(value, true, "Timestamp(micros/millis)"));
         mapping.put(
                 LogicalTypes.timestampMillis().getName(),
-                value ->
-                        AvroToProtoSerializerUtils.convertTimestamp(
-                                value, false, "Timestamp(micros/millis)"));
-        mapping.put(LogicalTypes.uuid().getName(), AvroToProtoSerializerUtils::convertUUID);
-        mapping.put(
-                LogicalTypes.timeMillis().getName(),
-                value -> AvroToProtoSerializerUtils.convertTime(value, false));
-        mapping.put(
-                LogicalTypes.timeMicros().getName(),
-                value -> AvroToProtoSerializerUtils.convertTime(value, true));
+                value -> convertTimestamp(value, false, "Timestamp(micros/millis)"));
+        mapping.put(LogicalTypes.uuid().getName(), AvroToProtoSerializer::convertUUID);
+        mapping.put(LogicalTypes.timeMillis().getName(), value -> convertTime(value, false));
+        mapping.put(LogicalTypes.timeMicros().getName(), value -> convertTime(value, true));
         mapping.put(
                 LogicalTypes.localTimestampMillis().getName(),
-                value -> AvroToProtoSerializerUtils.convertDateTime(value, false));
+                value -> convertDateTime(value, false));
         mapping.put(
                 LogicalTypes.localTimestampMicros().getName(),
-                value -> AvroToProtoSerializerUtils.convertDateTime(value, true));
-        mapping.put("geography_wkt", AvroToProtoSerializerUtils::convertGeography);
-        mapping.put("Json", AvroToProtoSerializerUtils::convertJson);
+                value -> convertDateTime(value, true));
+        mapping.put("geography_wkt", AvroToProtoSerializer::convertGeography);
+        mapping.put("Json", AvroToProtoSerializer::convertJson);
         return mapping.get(logicalTypeString);
+    }
+
+    // ---- Utilities to enable Conversions for Logical Types ------------
+    @VisibleForTesting
+    static String convertUUID(Object value) {
+        if (value instanceof UUID) {
+            return ((UUID) value).toString();
+        } else {
+            Preconditions.checkArgument(
+                    value instanceof String, "Expecting a value as String type.");
+            UUID.fromString((String) value);
+            return (String) value;
+        }
+    }
+
+    private static long validateTimestamp(long timestamp) {
+        // Since bigquery requires the timestamp to be in Microseconds since epoch.
+        // But UNIX considers it in Milliseconds since the epoch.
+        long timestampMillis = TimeUnit.MICROSECONDS.toMillis(timestamp);
+        Timestamp minTs = Timestamp.valueOf("0001-01-01 00:00:00");
+        Timestamp maxTs = Timestamp.valueOf("9999-12-31 23:59:59.999999");
+        Timestamp ts = null;
+        try {
+            ts = new Timestamp(timestampMillis);
+
+            if (ts.before(minTs) || ts.after(maxTs)) {
+                throw new IllegalArgumentException(
+                        String.format("Invalid Timestamp '%s' Provided", ts));
+            }
+            return timestamp;
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException(
+                    String.format(
+                            "Invalid Timestamp '%s' Provided."
+                                    + "\nShould be a long value indicating microseconds since Epoch (1970-01-01 00:00:00) "
+                                    + "between %s and %s",
+                            ts, minTs, maxTs));
+        }
+    }
+
+    // BigQuery inputs timestamp as microseconds since EPOCH,
+    // So if we have TIMESTAMP in micros - we convert as it is.
+    // If the TIMESTAMP is in millis - we convert to Micros and then add.
+    @VisibleForTesting
+    static Long convertTimestamp(Object value, boolean micros, String type) {
+        long timestamp;
+        if (value instanceof ReadableInstant) {
+            timestamp = TimeUnit.MILLISECONDS.toMicros(((ReadableInstant) value).getMillis());
+        } else {
+            Preconditions.checkArgument(
+                    value instanceof Long,
+                    String.format(
+                            "Expecting a value as Long type " + "%s. Instead %s was obtained",
+                            type, value.getClass()));
+            timestamp = (micros ? (Long) value : TimeUnit.MILLISECONDS.toMicros((Long) value));
+        }
+        return validateTimestamp(timestamp);
+    }
+
+    private static Integer validateDate(Integer date) {
+        if (date > 2932896 || date < -719162) {
+            throw new IllegalArgumentException(
+                    String.format(
+                            "Invalid Date '%s' Provided."
+                                    + "\nShould be a Integer value indicating days since Epoch (1970-01-01 00:00:00) "
+                                    + "between %s and %s",
+                            LocalDate.ofEpochDay(date), "0001-01-01", "9999-12-31"));
+        }
+        return date;
+    }
+
+    @VisibleForTesting
+    static Integer convertDate(Object value) {
+        // The value is the number of days since the Unix epoch (1970-01-01).
+        // The valid range is `-719162` (0001-01-01) to `2932896` (9999-12-31).
+        int date;
+        if (value instanceof ReadableInstant) {
+            date = Days.daysBetween(Instant.EPOCH, (ReadableInstant) value).getDays();
+        } else {
+            Preconditions.checkArgument(
+                    value instanceof Integer, "Expecting a value as Integer type (days).");
+            date = (Integer) value;
+        }
+        return validateDate(date);
+    }
+
+    @VisibleForTesting
+    static String convertDateTime(Object value, boolean micros) {
+        if (value instanceof String) {
+            return (String) value;
+        }
+        if (value instanceof Utf8) {
+            return ((Utf8) value).toString();
+        }
+        if (value instanceof Long) {
+            /*
+             * Assume that it is provided in {@link CivilTimeEncoder} Encoding Microsecond Precision
+             * format
+             */
+            return CivilTimeEncoder.decodePacked64DatetimeMicros((long) value).toString();
+        }
+        throw new UnsupportedOperationException(
+                String.format(
+                        "Expecting Local Timestamp(micros/millis) value in String/Long format.%nFound %s instead.",
+                        value.getClass()));
+    }
+
+    @VisibleForTesting
+    static String convertTime(Object value, boolean micros) {
+        if (value instanceof String) {
+            return (String) value;
+        }
+        if (value instanceof Long) {
+            /*
+             * Assume that it is provided in {@link CivilTimeEncoder} Encoding Microsecond Precision
+             * format
+             */
+            return CivilTimeEncoder.decodePacked64TimeMicros((long) value).toString();
+        }
+        throw new UnsupportedOperationException(
+                "Time(micros/millis) value not provided in String/Long format");
+    }
+
+    // 1. There is no way to check the precision and scale of NUMERIC/BIGNUMERIC fields,
+    // so we can just check by the maximum value.
+    // 2. decimal() logical type is mapped to BIGNUMERIC bigquery field.
+    // So in case we attempt to add decimal with precision >9 to a NUMERIC BQ field.
+    // The .append() method would be responsible for the error.
+    // .serialise() would successfully serialise it without any error indications.
+    @VisibleForTesting
+    static ByteString convertBigDecimal(Object value) {
+        // Assuming decimal (value) comes in big-endian encoding.
+        ByteBuffer byteBuffer = (ByteBuffer) value;
+        // Reverse before sending to big endian convertor.
+        // decodeBigNumericByteString() assumes string to be provided in little-endian.
+        byte[] byteArray = byteBuffer.array();
+        ArrayUtils.reverse(byteArray); // Converted to little-endian.
+        BigDecimal bigDecimal =
+                BigDecimalByteStringEncoder.decodeBigNumericByteString(
+                        ByteString.copyFrom(byteArray));
+        return BigDecimalByteStringEncoder.encodeToBigNumericByteString(bigDecimal);
+    }
+
+    @VisibleForTesting
+    static String convertGeography(Object value) {
+        if (value instanceof Utf8) {
+            return ((Utf8) value).toString();
+        }
+        Preconditions.checkArgument(
+                value instanceof String,
+                String.format(
+                        "Expecting a value as String/Utf8 type (geography_wkt or geojson format)."
+                                + "\nInstead got %s: ",
+                        value.getClass()));
+        return (String) value;
+    }
+
+    @VisibleForTesting
+    static String convertJson(Object value) {
+        String jsonString;
+        if (value instanceof Utf8) {
+            jsonString = ((Utf8) value).toString();
+        } else if (value instanceof String) {
+            jsonString = (String) value;
+        } else {
+            throw new IllegalArgumentException(
+                    String.format(
+                            "Expecting a value as String/Utf8 type (json format)."
+                                    + "\nInstead got %s: ",
+                            value.getClass()));
+        }
+        try {
+            new JSONObject(jsonString);
+        } catch (JSONException e) {
+            throw new IllegalArgumentException(
+                    String.format("The input string %s is not in valid JSON Format.", jsonString));
+        }
+        return jsonString;
     }
 }
